LayoutTests/fast/viewport/ios/viewport-fit-auto-safe-area-inset-change-expected.txt

Tests that changing the value of viewport-fit affects the size of the viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial width, using viewport-fit: cover
PASS document.body.clientWidth is viewportWidth

Set safe area, but still using viewport-fit: cover
PASS document.body.clientWidth is viewportWidth

Change to viewport-fit: auto
PASS document.body.clientWidth is viewportWidth - leftAndRightSafeAreaInsets

PASS successfullyParsed is true

TEST COMPLETE
